+++ pkg/server/setup.go
@@ -217,6 +217,7 @@ func (s *Server) setupCompleteHandler(w http.ResponseWriter, r *http.Request) {
 			Priority:       1,
 		}}
 		cfg.Usenet.MaxConnections = readerConnections
+		cfg.Usenet.ProcessingMaxConnections = readerConnections
 	} else {
 		cfg.Usenet.Providers = nil
 	}
@@ -240,7 +241,7 @@ func (s *Server) setupCompleteHandler(w http.ResponseWriter, r *http.Request) {
 		cfg.Categories = []string{"sonarr", "radarr"}
 	}
 	if cfg.MaxDownloads == 0 {
-		cfg.MaxDownloads = 10
+		cfg.MaxDownloads = 5
 	}
 	cfg.Mount.Type = config.MountType(req.Mount.MountType)